created a way to draw the food

+++ rocc/core/src/main/java/edu/chl/rocc/core/m2phyInterfaces/IFood.java
@@ -5,7 +5,7 @@ package edu.chl.rocc.core.m2phyInterfaces;
  */
 public interface IFood {
 
-    public int getX();
+    public float getX();
 
-    public int getY();
+    public float getY();
 }
